Extract profilePicFrame modifier
